Module aws_sdk_wisdom::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

CreateAssistant

Fluent builder constructing a request to CreateAssistant.

CreateAssistantAssociation

Fluent builder constructing a request to CreateAssistantAssociation.

CreateContent

Fluent builder constructing a request to CreateContent.

CreateKnowledgeBase

Fluent builder constructing a request to CreateKnowledgeBase.

CreateSession

Fluent builder constructing a request to CreateSession.

DeleteAssistant

Fluent builder constructing a request to DeleteAssistant.

DeleteAssistantAssociation

Fluent builder constructing a request to DeleteAssistantAssociation.

DeleteContent

Fluent builder constructing a request to DeleteContent.

DeleteKnowledgeBase

Fluent builder constructing a request to DeleteKnowledgeBase.

GetAssistant

Fluent builder constructing a request to GetAssistant.

GetAssistantAssociation

Fluent builder constructing a request to GetAssistantAssociation.

GetContent

Fluent builder constructing a request to GetContent.

GetContentSummary

Fluent builder constructing a request to GetContentSummary.

GetKnowledgeBase

Fluent builder constructing a request to GetKnowledgeBase.

GetRecommendations

Fluent builder constructing a request to GetRecommendations.

GetSession

Fluent builder constructing a request to GetSession.

ListAssistantAssociations

Fluent builder constructing a request to ListAssistantAssociations.

ListAssistants

Fluent builder constructing a request to ListAssistants.

ListContents

Fluent builder constructing a request to ListContents.

ListKnowledgeBases

Fluent builder constructing a request to ListKnowledgeBases.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

NotifyRecommendationsReceived

Fluent builder constructing a request to NotifyRecommendationsReceived.

QueryAssistant

Fluent builder constructing a request to QueryAssistant.

RemoveKnowledgeBaseTemplateUri

Fluent builder constructing a request to RemoveKnowledgeBaseTemplateUri.

SearchContent

Fluent builder constructing a request to SearchContent.

SearchSessions

Fluent builder constructing a request to SearchSessions.

StartContentUpload

Fluent builder constructing a request to StartContentUpload.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateContent

Fluent builder constructing a request to UpdateContent.

UpdateKnowledgeBaseTemplateUri

Fluent builder constructing a request to UpdateKnowledgeBaseTemplateUri.