Struct aws_sdk_wisdom::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for WisdomService.

This client allows ergonomic access to a WisdomService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_assistant(&self) -> CreateAssistant<C, M, R>

Constructs a fluent builder for the CreateAssistant operation.

See CreateAssistant for more information about the operation and its arguments.

pub fn create_assistant_association(
&self
) -> CreateAssistantAssociation<C, M, R>

Constructs a fluent builder for the CreateAssistantAssociation operation.

See CreateAssistantAssociation for more information about the operation and its arguments.

pub fn create_content(&self) -> CreateContent<C, M, R>

Constructs a fluent builder for the CreateContent operation.

See CreateContent for more information about the operation and its arguments.

pub fn create_knowledge_base(&self) -> CreateKnowledgeBase<C, M, R>

Constructs a fluent builder for the CreateKnowledgeBase operation.

See CreateKnowledgeBase for more information about the operation and its arguments.

pub fn create_session(&self) -> CreateSession<C, M, R>

Constructs a fluent builder for the CreateSession operation.

See CreateSession for more information about the operation and its arguments.

pub fn delete_assistant(&self) -> DeleteAssistant<C, M, R>

Constructs a fluent builder for the DeleteAssistant operation.

See DeleteAssistant for more information about the operation and its arguments.

pub fn delete_assistant_association(
&self
) -> DeleteAssistantAssociation<C, M, R>

Constructs a fluent builder for the DeleteAssistantAssociation operation.

See DeleteAssistantAssociation for more information about the operation and its arguments.

pub fn delete_content(&self) -> DeleteContent<C, M, R>

Constructs a fluent builder for the DeleteContent operation.

See DeleteContent for more information about the operation and its arguments.

pub fn delete_knowledge_base(&self) -> DeleteKnowledgeBase<C, M, R>

Constructs a fluent builder for the DeleteKnowledgeBase operation.

See DeleteKnowledgeBase for more information about the operation and its arguments.

pub fn get_assistant(&self) -> GetAssistant<C, M, R>

Constructs a fluent builder for the GetAssistant operation.

See GetAssistant for more information about the operation and its arguments.

pub fn get_assistant_association(&self) -> GetAssistantAssociation<C, M, R>

Constructs a fluent builder for the GetAssistantAssociation operation.

See GetAssistantAssociation for more information about the operation and its arguments.

pub fn get_content(&self) -> GetContent<C, M, R>

Constructs a fluent builder for the GetContent operation.

See GetContent for more information about the operation and its arguments.

pub fn get_content_summary(&self) -> GetContentSummary<C, M, R>

Constructs a fluent builder for the GetContentSummary operation.

See GetContentSummary for more information about the operation and its arguments.

pub fn get_knowledge_base(&self) -> GetKnowledgeBase<C, M, R>

Constructs a fluent builder for the GetKnowledgeBase operation.

See GetKnowledgeBase for more information about the operation and its arguments.

pub fn get_recommendations(&self) -> GetRecommendations<C, M, R>

Constructs a fluent builder for the GetRecommendations operation.

See GetRecommendations for more information about the operation and its arguments.

pub fn get_session(&self) -> GetSession<C, M, R>

Constructs a fluent builder for the GetSession operation.

See GetSession for more information about the operation and its arguments.

pub fn list_assistant_associations(&self) -> ListAssistantAssociations<C, M, R>

Constructs a fluent builder for the ListAssistantAssociations operation.

See ListAssistantAssociations for more information about the operation and its arguments.

pub fn list_assistants(&self) -> ListAssistants<C, M, R>

Constructs a fluent builder for the ListAssistants operation.

See ListAssistants for more information about the operation and its arguments.

pub fn list_contents(&self) -> ListContents<C, M, R>

Constructs a fluent builder for the ListContents operation.

See ListContents for more information about the operation and its arguments.

pub fn list_knowledge_bases(&self) -> ListKnowledgeBases<C, M, R>

Constructs a fluent builder for the ListKnowledgeBases operation.

See ListKnowledgeBases for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn notify_recommendations_received(
&self
) -> NotifyRecommendationsReceived<C, M, R>

Constructs a fluent builder for the NotifyRecommendationsReceived operation.

See NotifyRecommendationsReceived for more information about the operation and its arguments.

pub fn query_assistant(&self) -> QueryAssistant<C, M, R>

Constructs a fluent builder for the QueryAssistant operation.

See QueryAssistant for more information about the operation and its arguments.

pub fn remove_knowledge_base_template_uri(
&self
) -> RemoveKnowledgeBaseTemplateUri<C, M, R>

Constructs a fluent builder for the RemoveKnowledgeBaseTemplateUri operation.

See RemoveKnowledgeBaseTemplateUri for more information about the operation and its arguments.

pub fn search_content(&self) -> SearchContent<C, M, R>

Constructs a fluent builder for the SearchContent operation.

See SearchContent for more information about the operation and its arguments.

pub fn search_sessions(&self) -> SearchSessions<C, M, R>

Constructs a fluent builder for the SearchSessions operation.

See SearchSessions for more information about the operation and its arguments.

pub fn start_content_upload(&self) -> StartContentUpload<C, M, R>

Constructs a fluent builder for the StartContentUpload operation.

See StartContentUpload for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_content(&self) -> UpdateContent<C, M, R>

Constructs a fluent builder for the UpdateContent operation.

See UpdateContent for more information about the operation and its arguments.

pub fn update_knowledge_base_template_uri(
&self
) -> UpdateKnowledgeBaseTemplateUri<C, M, R>

Constructs a fluent builder for the UpdateKnowledgeBaseTemplateUri operation.

See UpdateKnowledgeBaseTemplateUri for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.