Module aws_sdk_wisdom::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

precondition_failed_exception

See PreconditionFailedException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient access to perform this action.

ConflictException

The request could not be processed because of conflict in the current state of the

CreateAssistantAssociationError

Error type for the CreateAssistantAssociation operation.

CreateAssistantError

Error type for the CreateAssistant operation.

CreateContentError

Error type for the CreateContent operation.

CreateKnowledgeBaseError

Error type for the CreateKnowledgeBase operation.

CreateSessionError

Error type for the CreateSession operation.

DeleteAssistantAssociationError

Error type for the DeleteAssistantAssociation operation.

DeleteAssistantError

Error type for the DeleteAssistant operation.

DeleteContentError

Error type for the DeleteContent operation.

DeleteKnowledgeBaseError

Error type for the DeleteKnowledgeBase operation.

GetAssistantAssociationError

Error type for the GetAssistantAssociation operation.

GetAssistantError

Error type for the GetAssistant operation.

GetContentError

Error type for the GetContent operation.

GetContentSummaryError

Error type for the GetContentSummary operation.

GetKnowledgeBaseError

Error type for the GetKnowledgeBase operation.

GetRecommendationsError

Error type for the GetRecommendations operation.

GetSessionError

Error type for the GetSession operation.

ListAssistantAssociationsError

Error type for the ListAssistantAssociations operation.

ListAssistantsError

Error type for the ListAssistants operation.

ListContentsError

Error type for the ListContents operation.

ListKnowledgeBasesError

Error type for the ListKnowledgeBases operation.

ListTagsForResourceError

Error type for the ListTagsForResource operation.

NotifyRecommendationsReceivedError

Error type for the NotifyRecommendationsReceived operation.

PreconditionFailedException

The provided revisionId does not match, indicating the content has been

QueryAssistantError

Error type for the QueryAssistant operation.

RemoveKnowledgeBaseTemplateUriError

Error type for the RemoveKnowledgeBaseTemplateUri operation.

ResourceNotFoundException

The specified resource does not exist.

SearchContentError

Error type for the SearchContent operation.

SearchSessionsError

Error type for the SearchSessions operation.

ServiceQuotaExceededException

You've exceeded your service quota. To perform the requested action, remove some of the

StartContentUploadError

Error type for the StartContentUpload operation.

TagResourceError

Error type for the TagResource operation.

TooManyTagsException

Amazon Connect Wisdom throws this exception if you have too many tags in your tag set.

UntagResourceError

Error type for the UntagResource operation.

UpdateContentError

Error type for the UpdateContent operation.

UpdateKnowledgeBaseTemplateUriError

Error type for the UpdateKnowledgeBaseTemplateUri operation.

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CreateAssistantAssociationErrorKind

Types of errors that can occur for the CreateAssistantAssociation operation.

CreateAssistantErrorKind

Types of errors that can occur for the CreateAssistant operation.

CreateContentErrorKind

Types of errors that can occur for the CreateContent operation.

CreateKnowledgeBaseErrorKind

Types of errors that can occur for the CreateKnowledgeBase operation.

CreateSessionErrorKind

Types of errors that can occur for the CreateSession operation.

DeleteAssistantAssociationErrorKind

Types of errors that can occur for the DeleteAssistantAssociation operation.

DeleteAssistantErrorKind

Types of errors that can occur for the DeleteAssistant operation.

DeleteContentErrorKind

Types of errors that can occur for the DeleteContent operation.

DeleteKnowledgeBaseErrorKind

Types of errors that can occur for the DeleteKnowledgeBase operation.

GetAssistantAssociationErrorKind

Types of errors that can occur for the GetAssistantAssociation operation.

GetAssistantErrorKind

Types of errors that can occur for the GetAssistant operation.

GetContentErrorKind

Types of errors that can occur for the GetContent operation.

GetContentSummaryErrorKind

Types of errors that can occur for the GetContentSummary operation.

GetKnowledgeBaseErrorKind

Types of errors that can occur for the GetKnowledgeBase operation.

GetRecommendationsErrorKind

Types of errors that can occur for the GetRecommendations operation.

GetSessionErrorKind

Types of errors that can occur for the GetSession operation.

ListAssistantAssociationsErrorKind

Types of errors that can occur for the ListAssistantAssociations operation.

ListAssistantsErrorKind

Types of errors that can occur for the ListAssistants operation.

ListContentsErrorKind

Types of errors that can occur for the ListContents operation.

ListKnowledgeBasesErrorKind

Types of errors that can occur for the ListKnowledgeBases operation.

ListTagsForResourceErrorKind

Types of errors that can occur for the ListTagsForResource operation.

NotifyRecommendationsReceivedErrorKind

Types of errors that can occur for the NotifyRecommendationsReceived operation.

QueryAssistantErrorKind

Types of errors that can occur for the QueryAssistant operation.

RemoveKnowledgeBaseTemplateUriErrorKind

Types of errors that can occur for the RemoveKnowledgeBaseTemplateUri operation.

SearchContentErrorKind

Types of errors that can occur for the SearchContent operation.

SearchSessionsErrorKind

Types of errors that can occur for the SearchSessions operation.

StartContentUploadErrorKind

Types of errors that can occur for the StartContentUpload operation.

TagResourceErrorKind

Types of errors that can occur for the TagResource operation.

UntagResourceErrorKind

Types of errors that can occur for the UntagResource operation.

UpdateContentErrorKind

Types of errors that can occur for the UpdateContent operation.

UpdateKnowledgeBaseTemplateUriErrorKind

Types of errors that can occur for the UpdateKnowledgeBaseTemplateUri operation.