Module aws_sdk_wisdom::model

Data structures used by operation inputs/outputs.

Modules

app_integrations_configuration

See AppIntegrationsConfiguration

assistant_association_data

See AssistantAssociationData

assistant_association_summary

See AssistantAssociationSummary

assistant_data

See AssistantData

assistant_summary

See AssistantSummary

content_data

See ContentData

content_reference

See ContentReference

content_summary

See ContentSummary

document

See Document

document_text

See DocumentText

filter

See Filter

highlight

See Highlight

knowledge_base_association_data

See KnowledgeBaseAssociationData

knowledge_base_data

See KnowledgeBaseData

knowledge_base_summary

See KnowledgeBaseSummary

notify_recommendations_received_error

See NotifyRecommendationsReceivedError

recommendation_data

See RecommendationData

rendering_configuration

See RenderingConfiguration

result_data

See ResultData

search_expression

See SearchExpression

server_side_encryption_configuration

See ServerSideEncryptionConfiguration

session_data

See SessionData

session_summary

See SessionSummary

Structs

AppIntegrationsConfiguration

Configuration information for Amazon AppIntegrations to automatically ingest content.

AssistantAssociationData

Information about the assistant association.

AssistantAssociationSummary

Summary information about the assistant association.

AssistantData

The assistant data.

AssistantSummary

Summary information about the assistant.

ContentData

Information about the content.

ContentReference

Reference information about the content.

ContentSummary

Summary information about the content.

Document

The document.

DocumentText

The text of the document.

Filter

A search filter.

Highlight

Offset specification to describe highlighting of document excerpts for rendering search

KnowledgeBaseAssociationData

Association information about the knowledge base.

KnowledgeBaseData

Information about the knowledge base.

KnowledgeBaseSummary

Summary information about the knowledge base.

NotifyRecommendationsReceivedError

An error occurred when creating a recommendation.

RecommendationData

Information about the recommendation.

RenderingConfiguration

Information about how to render the content.

ResultData

Information about the result.

SearchExpression

The search expression.

ServerSideEncryptionConfiguration

The KMS key used for encryption.

SessionData

Information about the session.

SessionSummary

Summary information about the session.

Enums

AssistantAssociationInputData

The data that is input into Wisdom as a result of the assistant association.

AssistantAssociationOutputData

The data that is output as a result of the assistant association.

AssistantStatus

AssistantType

AssociationType

ContentStatus

FilterField

FilterOperator

KnowledgeBaseStatus

KnowledgeBaseType

RelevanceLevel

SourceConfiguration

Configuration information about the external data source.